+++ generators/app/templates/gulpfile.js
@@ -9,7 +9,7 @@ var nodemon = require('nodemon');
 var argv = require('yargs').argv;
 
 var injectSources = ["./dist/web/scripts/**/*.js", './dist/web/assets/**/*.css']
-var typeScriptFiles = ["./src/**/*.ts"]
+var typeScriptFiles = ["./src/**/*.ts?"]
 var staticFiles = ["./src/app/**/*.html", "./src/app/**/*.ejs", "./src/app/web/assets/**/*"]
 var htmlFiles = ["./src/app/**/*.html", "./src/app/**/*.ejs"]
 var watcherfiles = ["./src/**/*.*"]
@@ -81,7 +81,7 @@ gulp.task('static:inject', ['static:copy'], function () {
     var injectOptions = {
         relative: false,
         ignorePath: 'dist/web',
-        addRootSlash: false
+        addRootSlash: true
     };
 
     return gulp.src(htmlFiles)