+++ generators/messageExtension/index.js
@@ -170,6 +170,7 @@ class MessageExtensionGenerator extends Generator {
                     type: 'input',
                     name: 'messageExtensionName',
                     message: 'What is the name of your Message Extension command?',
+                    default: this.options.title + ' Message Extension',
                     validate: (input) => {
                         return input.length > 0;
                     },
@@ -178,6 +179,7 @@ class MessageExtensionGenerator extends Generator {
                     type: 'input',
                     name: 'messageExtensionDescription',
                     message: 'Describe your Message Extension command?',
+                    default: `Description of ${this.options.title} message extension`,
                     validate: (input) => {
                         return input.length > 0;
                     }
@@ -203,6 +205,9 @@ class MessageExtensionGenerator extends Generator {
         if (this.options.messageExtension) {
             let manifestPath = "src/manifest/manifest.json";
             var manifest = this.fs.readJSON(manifestPath);
+            if (!manifest.messageExtensions) {
+                manifest.messageExtensions = [];
+            }
             manifest.messageExtensions.push({
                 botId: this.options.messageExtensionId,
                 scopes: ["team", "personal"],